Adds an alias for the AGE system of games
The AGE system games all use 3d6 for ability tests, with 1 of them typically being represented with a different color, size, markings, etc in meat space. This different die is referred to as the drama die, stunt die, or dragon die depending on the game, but it functions the same across all of them. For the AGE system it's important that all 3 dice be rolled together but that the drama die is able to be distinguished from the others. This can be done in Dice Maiden with `!roll 2d6 + 1d6`, but since _every single test_ is always made this way it would be helpful to have an alias for that so that players don't have to continuously type or copypasta the sequence. Games in the AGE system include Fantasy Age, Dragon Age, Modern Age, Blue Rose, and The Expanse RPG.