Resolve an issue with comments containing "dn"
So there was an issue with the wrath and glory difficulty number check. When using include? we were also grabbing any instance of "dn" in a comment which broke the roll. This update now uses match? and regex to check for the start of the input string only. No more kidney comments breaking rolls :)