Fix issue with e and ie syntax
When specifying explode or indefinite explode in your roll syntax, you should not have to always give a number. If no number is given, it is assumed that the explode target is the number of sides on the die. This was not working correctly since the math engine rework. This commit resolves that issue