Re add heartbeat check
In an attempt of stopping the infinite reconnect loop which causes the
bots api key to become revoked, I am re adding the heartbeat ack check
change. This will allow the bot to catch up if it misses a heartbeat and
not reconnect burning through the start quota. Some documentation can be
found here: https://github.com/discordrb/discordrb/issues/600

The bot may become slow to respond to rolls but it wont crash which is
good. Well hopefully...