Add alias and system help to bot
This lays the help text out more clearly in the code. Moving it from a single line to a multi-line string.

It also adds two more in-depth help options `help alias` which shows a list of the generic aliases available and `help system` which shows a list of the system specific help commands.

This should allow users to find the available commands a lot easier than having to go to GitHub and check the readme.