Return Alias Request in Output
Remove the comment from the input initially and then preserve the user request string before parsing aliases.

Then in the output use the initial request string rather than the parsed roll string.

This should make it easier for users to tell what they rolled in a meaningful way. The parsed roll string does not always relate clearly to the alias.