Fix rerolls and add private roll support
These changes resolve an issue where only the last part of a multi roll
is checked when it comes to rerolls.

This also adds support for private rolls. Now you can private message
the bot to hide both the roll syntax and the result but if that is too
much a hassle users can now do "/roll p 4d10" to get the result as a
private message in channel.

fixes #187
closes #120