Add Hero System
Thanks to @ThatGuyWhatsMadeofCheese for the heavy lifting here.
rewriting his original pull request to get it up to slash command stands
and fix a few errors.

This commit adds support for the hero system. This commit also adds
support for a no total mode that prints the tally but no the results.

resolves #46